Disable/remove the friend-finder. tangogps.org apparently no longer provides the friend-finder service, and we don't have a new back-end to replace that service yet, and trying to use the friend-finder feature with tangogps.org as it currently exists results in a crash.
Rather than debugging the crash to maintain the feature
in a non-functional state, the most reasonable thing to do
seems to be simply removing the feature for the time being.

The relevant widgets have been deleted from foxtrotgps.glade,
because we can dig them out of VC history later when we have
a replacement back-end; and the `fftimer_running' variable
is no longer read from GConf but rather just set to FALSE
so that the broken friend-finder doesn't try to auto-start.