Prevent logo/manual name-conflicts when making PDFs with newer Texinfo.
The texinfo.tex from Texinfo 4.12 (distributed with Automake 1.11.2)
tries to use PDF-formatted images in preference to PNG images,
and fails if there happens to be an incompletely-generated PDF
by the same name... because the manual itself has the same name
as one of the images used in it.

So, give the logo a distinct name of "foxtrotgps-logo".