Don't set toolbar icon-size.
This hasn't been working for a while, since libglade apparently changed the syntax it expects (GTK_ICON_SIZE_BUTTON just results in a error-message on the console and the defaults being used).

Since nobody's complained about the icons being the wrong size, we may as well just accept that the defaults (whatever they are) are correct.